requirements.txt
pandas
numpy
openai
python-dotenv
yfinance
